Recalculate class widths on remove
When a class was removed, the widths of all conflicting classes were not
being recalculated. This now fixes that, so all conflicting class widths
will readjust on class removal.

This was done through a refactor of all conflict/width related
calculation and color assignation (basically everything you need to do
to a section or an associated class to prep it for adding to the
schedule) being put into one function (prepClassesForCalendar) that can
be called when a class or an associated class is being added or removed.