Flash Rearchitecture
The usage during wildcat welcome far exceeded my predictions for the
firestore daily access limit, so I refactored firestore out. Data is now
stored in local static json files that will have to be committed along
with the code, just like it was in 2.0.

In hindsight, this is the only real solution that truly costs $0 even if
it comes with some manual work required for data updates.